model/requirements.txt
tensorflow>=2.20.0
keras>=3.0.0
numpy>=2.1.0
pillow>=10.4.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
